refactor: project structure with common go structure.
