+++ internal/clients/testdata/api_responses/charizard.json
